--- app/src/main/java/com/example/momolabfe/ui/record/CommonRecordInfoFragment.kt
@@ -1,261 +1,0 @@
-package com.example.momolabfe.ui.record
-
-import android.content.res.ColorStateList
-import android.graphics.Color
-import android.os.Bundle
-import android.text.Editable
-import android.util.Log
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.CompoundButton
-import androidx.core.content.ContextCompat
-import androidx.core.widget.addTextChangedListener
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.repeatOnLifecycle
-import com.example.momolabfe.R
-import com.example.momolabfe.data.remote.record.model.DayWeek
-import com.example.momolabfe.data.remote.record.model.RecordCreateRequest
-import com.example.momolabfe.data.remote.record.model.Turbidity
-import com.example.momolabfe.databinding.FragmentRecordCommonInfoBinding
-import com.example.momolabfe.ui.record.viewModel.RecordViewModel
-import com.example.momolabfe.utils.korean
-import com.example.momolabfe.utils.toDayWeek
-import kotlinx.coroutines.launch
-import com.google.android.material.bottomnavigation.BottomNavigationView
-import java.math.BigDecimal
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
-
-class CommonRecordInfoFragment : Fragment() {
-
-    private var _binding: FragmentRecordCommonInfoBinding? = null
-    private val binding get() = _binding!!
-
-    private val viewModel: RecordViewModel by activityViewModels()
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentRecordCommonInfoBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        // 바텀 내비게이션 숨기기
-        activity?.findViewById<BottomNavigationView>(R.id.main_bnv)?.visibility = View.GONE
-
-        // 전달된 날짜 문자열 꺼내기 (기본값: 오늘 날짜 포맷 or 빈 문자열)
-        val dateText = arguments?.getString("selected_date_text")
-            ?: LocalDate.now().format(DateTimeFormatter.ofPattern("yyyy년 M월 d일"))
-
-        val dwKorean = arguments?.getString("selected_date_dw")
-            ?.let { DayWeek.valueOf(it).korean() }
-            ?: LocalDate.now().dayOfWeek.toDayWeek().korean()
-
-        binding.dateTv.text = dateText
-        binding.dwTv.text = dwKorean
-
-        setupUI()
-        setupObservers()
-    }
-
-    private fun setupUI() {
-
-        // 텍스트 변경 감지 → 버튼 상태 갱신
-        val watcher: (Editable?) -> Unit = { updateNextButtonState() }
-        binding.weightEt.addTextChangedListener(afterTextChanged = watcher)
-        binding.systolicEt.addTextChangedListener(afterTextChanged = watcher)
-        binding.diastolicEt.addTextChangedListener(afterTextChanged = watcher)
-        binding.fastingGlucoseEt.addTextChangedListener(afterTextChanged = watcher)
-        binding.urineCountEt.addTextChangedListener(afterTextChanged = watcher)
-        binding.totalUfEt.addTextChangedListener(afterTextChanged = watcher)
-
-        // 체크박스 변경 감지
-        binding.turbidityNCheckbox.setOnCheckedChangeListener { button, isChecked ->
-            if (isChecked) {
-                binding.turbidityYCheckbox.isChecked = false
-                setCheckBoxTint(binding.turbidityYCheckbox, false)
-            }
-            setCheckBoxTint(button, isChecked)
-            updateNextButtonState()
-        }
-        binding.turbidityYCheckbox.setOnCheckedChangeListener { button, isChecked ->
-            if (isChecked) {
-                binding.turbidityNCheckbox.isChecked = false
-                setCheckBoxTint(binding.turbidityYCheckbox, false)
-            }
-            setCheckBoxTint(button, isChecked)
-            updateNextButtonState()
-        }
-
-        // 초기 진입 시 현재 체크 상태에 맞춰 틴트 정렬
-        listOf(binding.turbidityNCheckbox, binding.turbidityYCheckbox).forEach {
-            setCheckBoxTint(it, it.isChecked)
-        }
-
-        updateNextButtonState()
-
-        binding.nextBtn.setOnClickListener {
-            val request = createRecordCommonRequest()
-            viewModel.recordCommonByWriting(request)
-        }
-    }
-
-    private fun updateNextButtonState() {
-        val weight = binding.weightEt.text?.toString()?.trim().orEmpty()
-        val systolic = binding.systolicEt.text?.toString()?.trim().orEmpty()
-        val diastolic = binding.diastolicEt.text?.toString()?.trim().orEmpty()
-        val fastingGlucose = binding.fastingGlucoseEt.text?.toString()?.trim().orEmpty()
-        val urineCount = binding.urineCountEt.text?.toString()?.trim().orEmpty()
-        val totalUf = binding.totalUfEt.text?.toString()?.trim().orEmpty()
-
-        val turbidityChecked = binding.turbidityNCheckbox.isChecked || binding.turbidityYCheckbox.isChecked
-
-        val enabled = isValidWeight(weight) &&
-                isValidBpSys(systolic) &&
-                isValidBpDia(diastolic) &&
-                isValidBpRelation(systolic, diastolic) &&
-                isValidGlucose(fastingGlucose) &&
-                isValidUrineFreq(urineCount) &&
-                isValidTotalUf(totalUf) &&
-                turbidityChecked
-
-        binding.nextBtn.isEnabled = enabled
-        binding.nextBtn.setBackgroundColor(
-            if (enabled) requireContext().getColor(R.color.text_primary) else Color.parseColor("#F6F6F6")
-        )
-        binding.nextBtn.setTextColor(
-            if (enabled) requireContext().getColor(R.color.white) else requireContext().getColor(R.color.text_primary)
-        )
-    }
-
-    // 체중(kg): 20.0 ~ 300.0
-    private fun isValidWeight(s: String): Boolean {
-        val bd = s.toBigDecimalOrNull() ?: return false
-        return bd >= BigDecimal("20.0") && bd <= BigDecimal("300.0")
-    }
-
-    // 혈압 최고(수축): 70 ~ 240
-    private fun isValidBpSys(s: String): Boolean =
-        s.toIntOrNull()?.let { it in 70..240 } == true
-
-    // 혈압 최저(이완): 40 ~ 160
-    private fun isValidBpDia(s: String): Boolean =
-        s.toIntOrNull()?.let { it in 40..160 } == true
-
-    // 혈압 최고(수축) > 혈압 최저(이완)
-    private fun isValidBpRelation(sys: String, dia: String): Boolean {
-        val s = sys.toIntOrNull() ?: return false
-        val d = dia.toIntOrNull() ?: return false
-        return s > d
-    }
-
-    // 공복 혈당: 40 ~ 600
-    private fun isValidGlucose(s: String): Boolean =
-        s.toIntOrNull()?.let { it in 40..600 } == true
-
-    // 소변 횟수: 0 ~ 50
-    private fun isValidUrineFreq(s: String): Boolean =
-        s.toIntOrNull()?.let { it in 0..50 } == true
-
-    // 제수량 합계: -5000 ~ 5000
-    private fun isValidTotalUf(s: String): Boolean =
-        s.toIntOrNull()?.let { it in -5000..5000 } == true
-
-
-    private fun setCheckBoxTint(checkBox: CompoundButton, isChecked: Boolean) {
-        val color = ContextCompat.getColor(
-            requireContext(), if (isChecked) R.color.text_primary else R.color.gray
-        )
-        checkBox.buttonTintList = ColorStateList.valueOf(color)
-    }
-
-    private fun createRecordCommonRequest(): RecordCreateRequest {
-        val dateStr = binding.dateTv.text?.toString()?.trim().orEmpty()
-        val formatter = DateTimeFormatter.ofPattern("yyyy년 M월 d일")
-        val recordDate = LocalDate.parse(dateStr, formatter)
-
-        val recordDw: DayWeek = recordDate.dayOfWeek.toDayWeek()
-
-        val weight = binding.weightEt.text?.toString()?.trim()?.toDoubleOrNull()
-            ?: error("체중 파싱 에러")
-        val systolic = binding.systolicEt.text?.toString()?.trim()?.toIntOrNull()
-            ?: error("최고 혈압 파싱 에러")
-        val diastolic = binding.diastolicEt.text?.toString()?.trim()?.toIntOrNull()
-            ?: error("최저 혈압 파싱 에러")
-        val fastingGlucose = binding.fastingGlucoseEt.text?.toString()?.trim()?.toIntOrNull()
-            ?: error("공복 혈당 파싱 에러")
-        val urineCount = binding.urineCountEt.text?.toString()?.trim()?.toIntOrNull()
-            ?: error("소변 횟수 파싱 에러")
-        val totalUf = binding.totalUfEt.text?.toString()?.trim()?.toIntOrNull()
-            ?: error("제수량 합계 파싱 에러")
-
-        val turbidity: Turbidity = when {
-            binding.turbidityNCheckbox.isChecked -> Turbidity.NONE
-            binding.turbidityYCheckbox.isChecked -> Turbidity.PRESENT
-            else -> error("혼탁도가 선택되지 않았습니다.")
-        }
-
-        val notes = binding.notesEt.text?.toString()?.trim().takeUnless { it.isNullOrEmpty() }
-
-        return RecordCreateRequest(
-            recordDate = recordDate,
-            recordDw = recordDw,
-            weight = weight,
-            systolic = systolic,
-            diastolic = diastolic,
-            fastingGlucose = fastingGlucose,
-            urineCount = urineCount,
-            turbidity = turbidity,
-            totalUf = totalUf,
-            notes = notes,
-        )
-    }
-
-    private fun setupObservers() {
-
-        // 성공 이벤트는 Flow 수집으로 1회성 처리
-        viewLifecycleOwner.lifecycleScope.launch {
-            viewLifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
-                viewModel.recordCreated.collect { recId ->
-                    Log.d("RECORD_COMMON_BY_WRITING_FRAGMENT", "공통 정보 작성이 정상적으로 완료되었습니다.")
-
-                    val dateStr = binding.dateTv.text?.toString()?.trim().orEmpty()
-                    val formatter = DateTimeFormatter.ofPattern("yyyy년 M월 d일")
-                    val recordDate = LocalDate.parse(dateStr, formatter)
-                    val recordDw = recordDate.dayOfWeek.toDayWeek()
-
-                    val args = Bundle().apply {
-                        putLong("rec_id", recId)
-                        putString("record_date", recordDate.toString())
-                        putString("record_dw", recordDw.name)
-                        putString("record_date_text", dateStr)
-                        putString("record_dw_korean", binding.dwTv.text?.toString().orEmpty())
-                    }
-                    val fragment = RecordExchangeInfoFragment().apply { arguments = args }
-
-                    parentFragmentManager.beginTransaction()
-                        .replace(R.id.main_frm, fragment)
-                        .addToBackStack(null)
-                        .commit()
-                }
-            }
-        }
-
-        viewModel.errorMessage.observe(viewLifecycleOwner) { errorMsg ->
-            Log.e("RECORD_COMMON_BY_WRITING_FRAGMENT", errorMsg.toString())
-        }
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
