+++ app/src/main/java/com/example/momolabfe/ui/record/RecordWrite01Fragment.kt
@@ -1,0 +1,42 @@
+package com.example.momolabfe.ui.record
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.momolabfe.R
+import com.example.momolabfe.databinding.FragmentRecord01Binding
+import com.google.android.material.bottomnavigation.BottomNavigationView
+
+class Record01Fragment : Fragment() {
+
+    private var _binding: FragmentRecord01Binding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentRecord01Binding.inflate(inflater, container, false)
+
+        return binding.root
+    }
+
+    override fun onResume() {
+        super.onResume()
+        val bottomNav = activity?.findViewById<BottomNavigationView>(R.id.main_bnv)
+        bottomNav?.visibility = View.VISIBLE
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.nextBtn.setOnClickListener {
+            parentFragmentManager.beginTransaction()
+                .replace(R.id.main_frm, Record02Fragment())
+                .addToBackStack(null)
+                .commit()
+        }
+    }
+}
