--- app/src/main/res/drawable/ic_album_sv.xml
@@ -1,48 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:aapt="http://schemas.android.com/aapt"
-    android:width="21dp"
-    android:height="21dp"
-    android:viewportWidth="21"
-    android:viewportHeight="21">
-  <path
-      android:pathData="M4.5,0C3.307,0 2.162,0.474 1.318,1.318C0.474,2.162 0,3.307 0,4.5V16.5C0,17.694 0.474,18.838 1.318,19.682C2.162,20.526 3.307,21 4.5,21H16.5C17.694,21 18.838,20.526 19.682,19.682C20.526,18.838 21,17.694 21,16.5V4.5C21,3.307 20.526,2.162 19.682,1.318C18.838,0.474 17.694,0 16.5,0H4.5Z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:centerX="-8.25"
-          android:centerY="-11.812"
-          android:gradientRadius="58.069"
-          android:type="radial">
-        <item android:offset="0.338" android:color="#FF0FAFFF"/>
-        <item android:offset="0.529" android:color="#FF367AF2"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path
-      android:pathData="M19.683,19.682L11.826,11.827C11.652,11.653 11.445,11.514 11.217,11.42C10.99,11.326 10.746,11.277 10.5,11.277C10.254,11.277 10.01,11.326 9.782,11.42C9.555,11.514 9.348,11.653 9.174,11.827L1.32,19.681C2.163,20.525 3.307,21 4.5,21.001H16.5C17.091,21.001 17.676,20.884 18.222,20.658C18.769,20.432 19.265,20.1 19.683,19.682Z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:startX="7.876"
-          android:startY="11.278"
-          android:endX="10.072"
-          android:endY="21.515"
-          android:type="linear">
-        <item android:offset="0" android:color="#FFB3E0FF"/>
-        <item android:offset="1" android:color="#FF8CD0FF"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path
-      android:pathData="M14.25,9C15.493,9 16.5,7.993 16.5,6.75C16.5,5.507 15.493,4.5 14.25,4.5C13.007,4.5 12,5.507 12,6.75C12,7.993 13.007,9 14.25,9Z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:startX="13.35"
-          android:startY="4"
-          android:endX="14.994"
-          android:endY="9.918"
-          android:type="linear">
-        <item android:offset="0" android:color="#FFFDFDFD"/>
-        <item android:offset="1" android:color="#FFB3E0FF"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-</vector>
