--- app/src/main/res/drawable/ic_camera_sv.xml
@@ -1,76 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:aapt="http://schemas.android.com/aapt"
-    android:width="30dp"
-    android:height="30dp"
-    android:viewportWidth="30"
-    android:viewportHeight="30">
-  <path
-      android:pathData="M10.168,4.239C10.356,3.867 10.643,3.553 10.998,3.335C11.353,3.116 11.762,3 12.179,3H17.824C18.243,3 18.652,3.117 19.008,3.336C19.364,3.556 19.651,3.871 19.837,4.245L20.715,6H23.25C24.245,6 25.198,6.395 25.902,7.098C26.605,7.802 27,8.755 27,9.75V21.75C27,22.745 26.605,23.698 25.902,24.402C25.198,25.105 24.245,25.5 23.25,25.5H6.75C5.755,25.5 4.802,25.105 4.098,24.402C3.395,23.698 3,22.745 3,21.75V9.75C3,8.755 3.395,7.802 4.098,7.098C4.802,6.395 5.755,6 6.75,6H9.285L10.168,4.239Z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:centerX="-0.75"
-          android:centerY="4.5"
-          android:gradientRadius="35.631"
-          android:type="radial">
-        <item android:offset="0" android:color="#FFF08AF4"/>
-        <item android:offset="0.535" android:color="#FF9C6CFE"/>
-        <item android:offset="1" android:color="#FF4E44DB"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path
-      android:pathData="M10.168,4.239C10.356,3.867 10.643,3.553 10.998,3.335C11.353,3.116 11.762,3 12.179,3H17.824C18.243,3 18.652,3.117 19.008,3.336C19.364,3.556 19.651,3.871 19.837,4.245L20.715,6H23.25C24.245,6 25.198,6.395 25.902,7.098C26.605,7.802 27,8.755 27,9.75V21.75C27,22.745 26.605,23.698 25.902,24.402C25.198,25.105 24.245,25.5 23.25,25.5H6.75C5.755,25.5 4.802,25.105 4.098,24.402C3.395,23.698 3,22.745 3,21.75V9.75C3,8.755 3.395,7.802 4.098,7.098C4.802,6.395 5.755,6 6.75,6H9.285L10.168,4.239Z"
-      android:fillAlpha="0.5">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:centerX="18"
-          android:centerY="16.845"
-          android:gradientRadius="8.686"
-          android:type="radial">
-        <item android:offset="0" android:color="#FF312A9A"/>
-        <item android:offset="1" android:color="#00312A9A"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path
-      android:pathData="M20.993,15C20.993,16.591 20.361,18.117 19.235,19.243C18.11,20.368 16.584,21 14.993,21C13.401,21 11.875,20.368 10.75,19.243C9.625,18.117 8.993,16.591 8.993,15C8.993,13.409 9.625,11.883 10.75,10.757C11.875,9.632 13.401,9 14.993,9C16.584,9 18.11,9.632 19.235,10.757C20.361,11.883 20.993,13.409 20.993,15Z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:startX="11.25"
-          android:startY="9"
-          android:endX="17.25"
-          android:endY="23.25"
-          android:type="linear">
-        <item android:offset="0" android:color="#FFFFFFFF"/>
-        <item android:offset="1" android:color="#FFDECBFF"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path
-      android:pathData="M19.5,15C19.5,16.194 19.026,17.338 18.182,18.182C17.338,19.026 16.194,19.5 15,19.5C13.807,19.5 12.662,19.026 11.818,18.182C10.974,17.338 10.5,16.194 10.5,15C10.5,13.807 10.974,12.662 11.818,11.818C12.662,10.974 13.807,10.5 15,10.5C16.194,10.5 17.338,10.974 18.182,11.818C19.026,12.662 19.5,13.807 19.5,15Z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:centerX="9"
-          android:centerY="8.25"
-          android:gradientRadius="12.772"
-          android:type="radial">
-        <item android:offset="0.243" android:color="#FF3BD5FF"/>
-        <item android:offset="1" android:color="#FF2052CB"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path
-      android:pathData="M22.5,12C22.898,12 23.279,11.842 23.561,11.561C23.842,11.279 24,10.898 24,10.5C24,10.102 23.842,9.721 23.561,9.439C23.279,9.158 22.898,9 22.5,9C22.102,9 21.721,9.158 21.439,9.439C21.158,9.721 21,10.102 21,10.5C21,10.898 21.158,11.279 21.439,11.561C21.721,11.842 22.102,12 22.5,12Z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:startX="21"
-          android:startY="9.75"
-          android:endX="24"
-          android:endY="12"
-          android:type="linear">
-        <item android:offset="0" android:color="#FFF08AF4"/>
-        <item android:offset="1" android:color="#FFF462AB"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-</vector>
