--- app/src/main/res/drawable/ic_record_list_sv.xml
@@ -1,13 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="25dp"
-    android:height="25dp"
-    android:viewportWidth="25"
-    android:viewportHeight="25">
-  <path
-      android:pathData="M8.333,6.25H20.833M4.167,6.26L4.177,6.249M4.167,12.51L4.177,12.499M4.167,18.76L4.177,18.749M8.333,12.5H20.833M8.333,18.75H20.833"
-      android:strokeLineJoin="round"
-      android:strokeWidth="1.5625"
-      android:fillColor="#00000000"
-      android:strokeColor="#000000"
-      android:strokeLineCap="round"/>
-</vector>
