--- app/src/main/res/layout/fragment_record_select_date.xml
@@ -1,100 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ui.record.SelectRecordDateFragment">
-
-    <TextView
-        android:id="@+id/title_tv"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="120dp"
-        android:text="날짜를 선택해주세요"
-        android:textSize="25sp"
-        android:fontFamily="@font/noto_sans_kr_regular"
-        android:textColor="@color/text_primary"
-        android:textStyle="bold"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        android:layout_marginStart="20dp"
-        android:includeFontPadding="false"/>
-
-    <TextView
-        android:id="@+id/selected_date_tv"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="30dp"
-        android:fontFamily="@font/noto_sans_kr_regular"
-        android:gravity="center"
-        tools:text="2025년 10월"
-        android:textColor="@color/text_primary"
-        android:textSize="23sp"
-        android:textStyle="bold"
-        app:layout_constraintStart_toStartOf="@id/title_tv"
-        app:layout_constraintTop_toBottomOf="@id/title_tv" />
-
-    <ImageView
-        android:id="@+id/calendar_previous_date_iv"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="20dp"
-        android:src="@drawable/ic_calendar_previous_sv"
-        app:layout_constraintEnd_toStartOf="@id/calendar_next_date_iv"
-        app:layout_constraintTop_toTopOf="@id/selected_date_tv" />
-
-    <ImageView
-        android:id="@+id/calendar_next_date_iv"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="10dp"
-        android:src="@drawable/ic_calendar_next_sv"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="@id/calendar_previous_date_iv" />
-
-    <LinearLayout
-        android:id="@+id/calendar_weekdays_row"
-        android:layout_marginTop="20dp"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="10dp"
-        android:layout_marginEnd="10dp"
-        android:orientation="horizontal"
-        android:paddingBottom="8dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/selected_date_tv"/>
-
-    <com.kizitonwose.calendar.view.CalendarView
-        android:id="@+id/calender_view"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="10dp"
-        android:layout_marginEnd="10dp"
-        app:cv_dayViewResource="@layout/calendar_day_layout"
-        app:cv_orientation="horizontal"
-        app:cv_outDateStyle="endOfRow"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/calendar_weekdays_row" />
-
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/next_btn"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="다음으로"
-        android:textSize="15sp"
-        android:fontFamily="@font/noto_sans_kr_regular"
-        android:textColor="@color/white"
-        app:cornerRadius="12dp"
-        android:backgroundTint="@color/text_primary"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        android:layout_marginStart="19dp"
-        android:layout_marginEnd="19dp"
-        android:layout_marginBottom="50dp"/>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
