+++ app/src/main/res/layout/fragment_record_write_01.xml
@@ -3,15 +3,14 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.record.CommonRecordInfoFragment">
+    android:layout_height="match_parent">
 
     <androidx.core.widget.NestedScrollView
         android:id="@+id/scroll_container"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:clipToPadding="false"
-        android:paddingBottom="140dp">
+        android:paddingBottom="50dp">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
@@ -186,7 +185,7 @@
 
                     <androidx.cardview.widget.CardView
                         android:id="@+id/systolic_cv"
-                        android:layout_width="147dp"
+                        android:layout_width="140dp"
                         android:layout_height="wrap_content"
                         android:layout_marginTop="16dp"
                         android:backgroundTint="@color/edit_card"
@@ -231,7 +230,7 @@
 
                     <androidx.cardview.widget.CardView
                         android:id="@+id/diastolic_cv"
-                        android:layout_width="147dp"
+                        android:layout_width="140dp"
                         android:layout_height="wrap_content"
                         android:layout_marginStart="10dp"
                         android:layout_marginTop="16dp"
@@ -476,7 +475,7 @@
                         android:layout_height="wrap_content"
                         android:layout_marginTop="16dp"
                         android:layout_marginEnd="20dp"
-                        android:layout_marginBottom="20dp"
+                        android:layout_marginBottom="30dp"
                         android:backgroundTint="@color/edit_card"
                         app:cardCornerRadius="8dp"
                         app:cardElevation="0dp"
@@ -516,9 +515,9 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="19dp"
-                android:layout_marginTop="10dp"
+                android:layout_marginTop="20dp"
                 android:layout_marginEnd="19dp"
-                android:layout_marginBottom="50dp"
+                android:layout_marginBottom="20dp"
                 android:backgroundTint="#2B7FFF"
                 android:fontFamily="@font/noto_sans_kr_regular"
                 android:text="다음: 회차별 정보 입력"
