+++ app/src/main/res/layout/fragment_record_write_02.xml
